pyproject.toml
[project]
name = "kusogaki-bot"
version = "0.8.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts>=0.2.1",
    "deptry>=0.23.0",
    "discord-py>=2.4.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.37",
    "watchdog>=6.0.0",
    "dnspython>=2.7.0",
    "pymongo>=4.13.2"
]

[dependency-groups]
dev = [
    "pre-commit>=4.1.0",
    "ruff>=0.9.4",
]

[tool.deptry]
ignore = ["DEP001", "DEP003"]

[tool.deptry.per_rule_ignores]
DEP002 = ["psycopg2-binary", "audioop-lts", "deptry"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["kusogaki_bot"]